AI: refactored the packgage name
