+++ bin/src/test/java/com/qa/hwa/HwaSwimrApplicationTests.class
