--- src/main/java/com/example/hwa/HwaSwimrApplication.java
@@ -1,14 +1,0 @@
-package com.example.hwa;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HwaSwimrApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HwaSwimrApplication.class, args);
-		System.out.println("Swimr running");
-	}
-
-}
