--- src/main/java/com/example/hwa/domain/Club.java
@@ -1,110 +1,0 @@
-package com.example.hwa.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-
-@Entity
-public class Club {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-	private String name;
-	@OneToMany(mappedBy = "club")
-	private List<Place> places = new ArrayList<>();
-	
-	public Club() {};
-	
-	public Club(Integer id, String name) {
-		super();
-		this.id = id;
-		this.name = name;
-	}
-	
-	public Club(String name) {
-		super();
-		this.name = name;
-	}
-	
-	public Club(String name, List<Place> places) {
-		super();
-		this.name = name;
-		this.places = places;
-	}
-
-	public Club(Integer id, String name, List<Place> places) {
-		super();
-		this.id = id;
-		this.name = name;
-		this.places = places;
-	}
-
-	public Integer getId() {
-		return id;
-	}
-	public void setId(Integer id) {
-		this.id = id;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public List<Place> getPlaces() {
-		return places;
-	}
-	public void setPlaces(List<Place> places) {
-		this.places = places;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		result = prime * result + ((name == null) ? 0 : name.hashCode());
-		result = prime * result + ((places == null) ? 0 : places.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Club other = (Club) obj;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		if (name == null) {
-			if (other.name != null)
-				return false;
-		} else if (!name.equals(other.name))
-			return false;
-		if (places == null) {
-			if (other.places != null)
-				return false;
-		} else if (!places.equals(other.places))
-			return false;
-		return true;
-	}
-	
-	
-	
-	
-	
-}
