--- src/main/java/com/example/hwa/utils/Mapper.java
@@ -1,6 +1,0 @@
-package com.example.hwa.utils;
-
-public interface Mapper<E, D> {
-	D mapToDTO(E entity);
-	E mapFromDTO(D dto);
-}
