+++ src/main/java/com/qa/hwa/HwaSwimrApplication.java
@@ -1,0 +1,14 @@
+package com.qa.hwa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HwaSwimrApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HwaSwimrApplication.class, args);
+		System.out.println("Swimr running");
+	}
+
+}
