+++ src/main/java/com/qa/hwa/domain/Place.java
@@ -1,0 +1,128 @@
+package com.qa.hwa.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+@Entity
+public class Place {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Integer id;
+	private String name;
+	private String postcode;
+	@ManyToOne
+	private Club club;
+	
+	public Place() {};
+	
+	public Place(Integer id, String name, String postcode) {
+		super();
+		this.id = id;
+		this.name = name;
+		this.postcode = postcode;
+	}
+	
+	public Place(String name, String postcode) {
+		super();
+		this.name = name;
+		this.postcode = postcode;
+	}
+	
+//	public Place(String name, String postcode, Club club) {
+//		super();
+//		this.name = name;
+//		this.postcode = postcode;
+//		this.club = club;
+//	}
+
+	public Place(Integer id, String name, String postcode, Club club) {
+		super();
+		this.id = id;
+		this.name = name;
+		this.postcode = postcode;
+		this.club = club;
+	}
+
+	public Integer getId() {
+		return id;
+	}
+	public void setId(Integer id) {
+		this.id = id;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public String getPostcode() {
+		return postcode;
+	}
+	public void setPostcode(String postcode) {
+		this.postcode = postcode;
+	}
+	public Club getClub() {
+		return club;
+	}
+	public void setClub(Club club) {
+		this.club = club;
+	}
+
+//	@Override
+//	public int hashCode() {
+//		final int prime = 31;
+//		int result = 1;
+//		result = prime * result + ((club == null) ? 0 : club.hashCode());
+//		result = prime * result + ((id == null) ? 0 : id.hashCode());
+//		result = prime * result + ((name == null) ? 0 : name.hashCode());
+//		result = prime * result + ((postcode == null) ? 0 : postcode.hashCode());
+//		return result;
+//	}
+//
+//	@Override
+//	public boolean equals(Object obj) {
+//		if (this == obj)
+//			return true;
+//		if (obj == null)
+//			return false;
+//		if (getClass() != obj.getClass())
+//			return false;
+//		Place other = (Place) obj;
+//		if (club == null) {
+//			if (other.club != null)
+//				return false;
+//		} else if (!club.equals(other.club))
+//			return false;
+//		if (id == null) {
+//			if (other.id != null)
+//				return false;
+//		} else if (!id.equals(other.id))
+//			return false;
+//		if (name == null) {
+//			if (other.name != null)
+//				return false;
+//		} else if (!name.equals(other.name))
+//			return false;
+//		if (postcode == null) {
+//			if (other.postcode != null)
+//				return false;
+//		} else if (!postcode.equals(other.postcode))
+//			return false;
+//		return true;
+//	}
+//
+//	@Override
+//	public String toString() {
+//		return "Place [id=" + id + ", name=" + name + ", postcode=" + postcode + ", club=" + club + "]";
+//	}
+	
+	
+	
+	
+	
+}
+
+
