+++ src/main/java/com/qa/hwa/utils/Mapper.java
@@ -1,0 +1,6 @@
+package com.qa.hwa.utils;
+
+public interface Mapper<E, D> {
+	D mapToDTO(E entity);
+	E mapFromDTO(D dto);
+}
