+++ src/main/resources/application-test.properties
@@ -1,4 +1,5 @@
 #local h2 db
+server.port = 8081
 spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.username=sa
 spring.datasource.password=
