--- src/test/java/com/example/hwa/HwaSwimrApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.hwa;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HwaSwimrApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
