--- src/test/java/com/example/hwa/acceptance/IndexPageTest.java
@@ -1,33 +1,0 @@
-package com.example.hwa.acceptance;
-import static org.junit.Assert.assertEquals;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.Dimension;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class IndexPageTest {
-	 private WebDriver driver;
-
-//	    @Before
-//	    public void setup() {
-//	        System.setProperty("webdriver.chrome.driver", "src/test/resources/chromedriver.exe");
-//	        driver = new ChromeDriver();
-//	        driver.manage().window().setSize(new Dimension(1000, 768));
-//
-//	    }
-//
-//	    @Test
-//	    public void test() throws InterruptedException {
-//	        driver.get("https://localhost:8080/index.html");
-//
-//	        assertEquals("Swimr", driver.getTitle());
-//	    }
-//
-//	    @After
-//	    public void tearDown() {
-//	        driver.close();
-//	    }
-}
