--- src/test/java/com/example/hwa/acceptance/PlacePageTest.java
@@ -1,5 +1,0 @@
-package com.example.hwa.acceptance;
-
-public class PlacePageTest {
-
-}
