+++ src/test/java/com/qa/hwa/HwaSwimrApplicationTests.java
@@ -1,0 +1,13 @@
+package com.qa.hwa;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HwaSwimrApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
