+++ src/test/java/com/qa/hwa/acceptance/PlacePageTest.java
@@ -1,0 +1,8 @@
+package com.qa.hwa.acceptance;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class PlacePageTest {
+ 
+}
